#1453 test: move generic integration test to its own class
